Moved `http` mocks that are only used in that package
