+++ mock/http_mocks.go
@@ -5,58 +5,8 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/kkrull/gohttp/http"
 	. "github.com/onsi/gomega"
 )
-
-type Handler struct {
-	handleRequestReader  *bufio.Reader
-	handleResponseWriter io.Writer
-}
-
-func (mock *Handler) Handle(requestReader *bufio.Reader, responseWriter io.Writer) {
-	mock.handleRequestReader = requestReader
-	mock.handleResponseWriter = responseWriter
-}
-
-func (mock *Handler) ShouldHandleConnection() {
-	Expect(mock.handleRequestReader).NotTo(BeNil())
-	Expect(mock.handleResponseWriter).NotTo(BeNil())
-}
-
-func (mock *Handler) Routes() []http.Route {
-	return nil
-}
-
-type Router struct {
-	ReturnsRequest http.Request
-	ReturnsError   http.Response
-	receivedReader *bufio.Reader
-	parsed         []byte
-}
-
-func (mock *Router) Routes() []http.Route {
-	return nil
-}
-
-func (mock *Router) RouteRequest(reader *bufio.Reader) (http.Request, http.Response) {
-	mock.receivedReader = reader
-	allButLF, _ := reader.ReadBytes(byte('\r'))
-	shouldBeLF, _ := reader.ReadByte()
-	mock.parsed = appendByte(allButLF, shouldBeLF)
-	return mock.ReturnsRequest, mock.ReturnsError
-}
-
-func appendByte(allButLast []byte, last byte) []byte {
-	whole := make([]byte, len(allButLast)+1)
-	copy(whole, allButLast)
-	whole[len(whole)-1] = last
-	return whole
-}
-
-func (mock Router) VerifyReceived(reader *bufio.Reader) {
-	Expect(mock.receivedReader).To(BeIdenticalTo(reader))
-}
 
 type Request struct {
 	HandleReturns  string
@@ -74,42 +24,4 @@ func (mock *Request) Handle(writer io.Writer) error {
 
 func (mock *Request) VerifyHandle(writer *bufio.Writer) {
 	ExpectWithOffset(1, mock.handleReceived).To(BeIdenticalTo(writer))
-}
-
-type Route struct {
-	RouteReturns   http.Request
-	routeRequested http.RequestMessage
-}
-
-func (mock *Route) Route(requested http.RequestMessage) http.Request {
-	mock.routeRequested = requested
-	return mock.RouteReturns
-}
-
-func (mock *Route) ShouldHaveReceived(method string, target string) {
-	ExpectWithOffset(1, mock.routeRequested.Method()).To(Equal(method))
-	ExpectWithOffset(1, mock.routeRequested.Target()).To(Equal(target))
-}
-
-func (mock *Route) ShouldHaveReceivedParameters(parameters map[string]string) {
-	ExpectWithOffset(1, mock.routeRequested.QueryParameters()).To(Equal(parameters))
-}
-
-type Response struct {
-	writeHeaderReceived io.Writer
-	writtenTo           io.Writer
-}
-
-func (mock *Response) WriteTo(client io.Writer) error {
-	mock.writtenTo = client
-	return nil
-}
-
-func (mock *Response) WriteHeader(client io.Writer) error {
-	mock.writeHeaderReceived = client
-	return nil
-}
-
-func (mock *Response) VerifyWrittenTo(writer *bufio.Writer) {
-	ExpectWithOffset(1, mock.writtenTo).To(BeIdenticalTo(writer))
 }
